+++ changelog.md
@@ -7,6 +7,7 @@ Make sure to check it out at https://discord.gg/VrDkvDV4
 
 Support for Glowing Action Buttons (GABs), making some actions glow, namely:
 - Hunter's Aimed/Arcane/Chimera Shot buttons glow during Improved Steady Shot
+- Hunter's Arcane/Explosive Shot buttons glow during Lock and Load
 - Mage's Arcane Missiles button glows during Missile Barrage
 - Mage's Pyroblast button glows during Hot Streak
 - Mage's Fire Blast button glows during Impact